added fixes to resolve cache issues when download again
When downloading from the same table, after making changes to table (
loading new data or adding new lines ), it would keep using the same
exact file because of caching.  So set the stream cache to 0 to turn off
this caching.

Also added some better file naming conventions.  If CSVExporter, and
download file name ends in ".csv", use as is ( no change ).  Else, add
date/time stamp to make file name unique.  Same with ExcelExporter, but
use file name as is if it ends in ".xls" ( no change ), adding  time
stamp otherwise.  In both cases, in `Exporter.setContainerToBeExported(
container );`, we call `setDownloadFileName(getDownloadFileName);` to
refresh the file name automatically when data is loaded into the
exporter.

I have found that in Chrome, if I set the file name for "Excel", such as
"demo-excel.xls", it would keep using the cache.  As such, best case is
to use "demo-excel" ( without a file suffix ), seems to be more
realiable.